refactored solvers
