+++ solvers/adam.py
@@ -1,9 +1,9 @@
-from benchopt import BaseSolver
-from pytorch_lightning import Trainer
+from benchopt import safe_import_context
 from torch.optim import Adam
+with safe_import_context() as import_ctx:
+    TorchSolver = import_ctx.import_from('torch_solver', 'TorchSolver')
 
-
-class Solver(BaseSolver):
+class Solver(TorchSolver):
     """Adam solver"""
     name = 'Adam'
 
@@ -18,18 +18,8 @@ class Solver(BaseSolver):
         return False, None
 
     def set_objective(self, pl_module, trainer):
-        self.pl_module = pl_module
-        self.main_trainer = trainer  # we use this in order
-        # to access some elements from the trainer when
-        # initializing it below
+        super().set_objective(pl_module, trainer)
         self.pl_module.configure_optimizers = lambda: Adam(
             self.pl_module.parameters(),
             lr=self.lr,
         )
-
-    def run(self, n_iter):
-        trainer = Trainer(max_steps=n_iter)
-        trainer.fit(self.pl_module)
-
-    def get_result(self):
-        return self.pl_module
