+++ solvers/sgd.py
@@ -1,9 +1,9 @@
-from benchopt import BaseSolver
-from pytorch_lightning import Trainer
+from benchopt import safe_import_context
 from torch.optim import SGD
+with safe_import_context() as import_ctx:
+    TorchSolver = import_ctx.import_from('torch_solver', 'TorchSolver')
 
-
-class Solver(BaseSolver):
+class Solver(TorchSolver):
     """Stochastic Gradient descent solver"""
     name = 'SGD'
 
@@ -19,22 +19,13 @@ class Solver(BaseSolver):
     def skip(self, pl_module, trainer):
         if not self.momentum and self.nesterov:
             return True, 'Nesterov cannot be used without momentum'
+        return False, None
 
     def set_objective(self, pl_module, trainer):
-        self.pl_module = pl_module
-        self.main_trainer = trainer  # we use this in order
-        # to access some elements from the trainer when3
-        # initializing it below
+        super().set_objective(pl_module, trainer)
         self.pl_module.configure_optimizers = lambda: SGD(
             self.pl_module.parameters(),
             lr=self.lr,
             momentum=self.momentum,
             nesterov=self.nesterov,
         )
-
-    def run(self, n_iter):
-        trainer = Trainer(max_steps=n_iter)
-        trainer.fit(self.pl_module)
-
-    def get_result(self):
-        return self.pl_module
